polishing the sample code
