+++ examples/quick_start_00.rs
@@ -1,3 +1,7 @@
+// (C) 2025 - Enzo Lombardi
+//
+// CTRL+C or ALt+X or ESC-ESC to exit.
+
 use turbo_vision::prelude::*;
 fn main() -> turbo_vision::core::error::Result<()> {
     let mut app = Application::new()?;
