+++ examples/quick_start_01.rs
@@ -1,4 +1,5 @@
 // (C) 2025 - Enzo Lombardi
+// Add a status line to the desktop
 
 use turbo_vision::core::event::{KB_ALT_X, KB_ESC, KB_ESC_ESC};
 use turbo_vision::prelude::*;
